requirements.txt
pykml==0.2.0
skyfield==1.46
time-machine==2.10.0
flask==2.2.2
Werkzeug==2.2.2
jinja2==3.1.2
flask_login==0.6.2
flask_migrate==3.1.0
WTForms==3.0.1
flask_wtf==1.0.1
flask_sqlalchemy==2.5.1
sqlalchemy==1.4.23
email_validator==1.3.0
gunicorn==20.1.0
flask-restx==0.5.1
python-dotenv==0.21.0
Flask-Minify==0.39
elasticsearch==6.4
#elasticsearch[async]==8.4.2
python-dateutil==2.8.2
jira==3.4.1
beautifulsoup4==4.11.1
schedule==1.1.0
Flask-Caching==2.0.1
redis==4.3.4
czml==0.3.3
satellite-czml==0.1.2
satellitetle==0.14.0